Data.frame of base information for all ECOS species

Data.frame of base information for all ECOS species

TECP_table

Format

A data frame with 2445 rows and 10 variables

Scientific_Name

The scientific name, as recorded in ECOS

Common_Name

The common name, as recorded in ECOS

Species_Code

The four-character code assigned to species in ECOS

Critical_Habitat

CFR section under which CH was declared

Species_Group

Taxonomic group of species, as recorded in ECOS

Lead_Region

FWS region responsible for recovery

Federal_Listing_Status

At time of scraping

Special_Rules

CFR section under which any special rules were made

U_S__or_ForeignListed

One of US, US/Foreign, Foreign

Where_Listed

Geographic extent of listed entity

Species_Page

URL dedicated to the species on ECOS

Source

http://ecos.fws.gov/tess_public

Details

The data is the table returned using the URL:

https://goo.gl/tZRu5o

By default, ecosscraper checks this URL .onAttach to ensure - assuming the user restarts their sessions regularly - that the source is up-to-date. If the URL throws a http_error then this built-in data.frame will be loaded.

Note

Also notice that ecosscraper includes get_TECP_table, which can be used to 'override' the built-in data.

The source URL can be changed using options()$TE_list. The current data was scraped on 29 Nov 2016

See also

get_TECP_table