Get a df of threatened, endangered, candidate, and proposed species from ECOS

Returns threatened, endangered, candidate, and proposed species from ECOS.

get_TECP_table()

Value

A data.frame with 11 variables:

Scientific_Name

Common_Name

Species_Code

Critical_Habitat

Species_Group

Lead_Region

Federal_Listing_Status

Special_Rules

U_S__or_ForeignListed

Where_Listed

Species_Page

Details

Uses the URL set in options()$TE_list to get a data.frame that includes links for fetching data for any species with pages on TESS. This data is loaded

Examples

# To replace the TECP_table that is loaded on startup (attach)
TECP_table <- get_TECP_table()
head(TECP_table)

#> # A tibble: 6 × 11
#>           Scientific_Name                Common_Name Species_Code
#>                     <chr>                      <chr>        <chr>
#> 1     Abies guatemalensis Guatemalan Fir (=pinabete)         R003
#> 2      Abronia macrocarpa Large-fruited sand-verbena         Q24E
#> 3 Abutilon eremitopetalum             No common name         Q1SJ
#> 4      Abutilon menziesii                Ko`oloa`ula         Q1SK
#> 5    Abutilon sandwicense             No common name         Q1SL
#> 6           Acaena exigua                    Liliwai         Q1SM
#> # ... with 8 more variables: Critical_Habitat <chr>, Species_Group <chr>,
#> #   Lead_Region <chr>, Federal_Listing_Status <chr>, Special_Rules <chr>,
#> #   U.S._or_ForeignListed <chr>, Where_Listed <chr>, Species_Page <chr>