Gather all links for all or a subset of species on ECOS

Collects all of the links (URLs) from ECOS for a set of species.

get_bulk_species_links(urls, parallel = TRUE)

Arguments

urls

A vector of urls from ECOS to visit for link-gathering

parallel

Try parallel scraping [default = TRUE]

Value

A data.frame with variables:

Scientific_Name

Self-explanatory

href

The URL as given on the ECOS page

link

The URL with domain information added

text

The text representation of the URL

Details

Automates the process of gathering links, which may in turn be followed or, in the case of documents or tables, be downloaded. This is a convenience wrapper over get_species_links.

See also

remove_silly_links get_species_links

Examples

## Not run: ------------------------------------
#   res <- get_bulk_species_links(TECP_table$Species_Page[1:3])
## ---------------------------------------------